Finally got colored output and cleaner code by taking a different route
Basically got rid of using Named Pipes to get this to work. They didn't help, made the code more complex, did not allow for input and output on the same pipe, and could not give me color output, so I dropped it. Instead I've taken the route of just "forking" the process, calling some internal flags and passing it the parameters the original executable was called with. Then I free all Standard I/O handles inherited on the child process and attach it to the parent. This, which not only allows me to get colored output, but console input functions properly without any extra work or steps taken.